fuse: fixes for RHEL10
Fuse in the linux kernel changed slightly since the patches to
remove tmp page copies were developed for linux 6.8.
The main logic did not change only some convenience functions
were added to the kernel which the patch did not use for
compatibility reasons since they were not present in lunx 6.8.

- use fuse_writepage_finish_stat() where appropriate
- use fuse_writepage_args_setup() to setup the args

Signed-off-by: Horst Birthelmer <hbirthelmer@ddn.com>