fuse: remove tmp_folios in the launder folios path to avoid deadlock
when calling
	fuse_launder_folio()
		->fuse_writepage_locked()
			->fuse_flush_writepages(inode)

The system waits for all pages to be written back.
For the tmp folios created in this path this would never happen
since the previous patch removed dependency on NR_WRITEBACK_TEMP.

note that this is a side effect of the removal of tmp pages patch
and not a general problem.